+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/ElectricFurnace.java
@@ -1,15 +1,12 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.electric.machines;
 
-import java.util.Iterator;
-
-import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.inventory.FurnaceRecipe;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.Recipe;
 import org.bukkit.inventory.RecipeChoice;
 import org.bukkit.inventory.RecipeChoice.MaterialChoice;
 
+import me.mrCookieSlime.Slimefun.SlimefunPlugin;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.AContainer;
@@ -23,20 +20,17 @@ public abstract class ElectricFurnace extends AContainer {
 
     @Override
     public void registerDefaultRecipes() {
-        Iterator<Recipe> iterator = Bukkit.recipeIterator();
-        while (iterator.hasNext()) {
-            Recipe recipe = iterator.next();
-            
-            if (recipe instanceof FurnaceRecipe) {
-                RecipeChoice choice = ((FurnaceRecipe) recipe).getInputChoice();
-                
+        SlimefunPlugin.getMinecraftRecipes().subscribe(snapshot -> {
+            for (FurnaceRecipe recipe : snapshot.getRecipes(FurnaceRecipe.class)) {
+                RecipeChoice choice = recipe.getInputChoice();
+
                 if (choice instanceof MaterialChoice) {
                     for (Material input : ((MaterialChoice) choice).getChoices()) {
                         registerRecipe(4, new ItemStack[] { new ItemStack(input) }, new ItemStack[] { recipe.getResult() });
                     }
                 }
             }
-        }
+        });
     }
 
     @Override